requirements.txt
transformers
pillow
torch
torchvision
fastapi
uvicorn
python-multipart
huggingface_hub
numpy
safetensors
accelerate
scipy
timm
kornia
flet==0.25.2
